refactor(kubernetes): simplify Kubernetes credentials classes (#4793)
* refactor(kubernetes): remove KubernetesCredentials interface

There is now only one implementation of the KubernetesCredentials interface: KubernetesV2Credentials; let's remove KubernetesCredentials to simplify the code.

Most of this diff is removing type parameters from KubernetesCredentialFactory and KubernetesNamedAccountCredentials, as these are no longer necessary.

* refactor(kubernetes): remove KubernetesCredentialsFactory interface

The KubernetesCredentialsFactory has a single implementation, KubernetesV2Credentials.Factory. Move the default getKubeconfigFile method to this implementation and remove the interface. Replace the stub implementation used only by the KubernetesHealthIndicatorTest with mocks in the test itself.

* refactor(kubernetes): rename KubernetesV2Credentials to KubernetesCredentials

* fix(kubernetes): fix KubernetesHealthIndicatorTest

The commit that removed the StubKubernetesCredentialsFactory broke this test because it tried to create a mocked item in a `thenReturn` statement of another mock. I was finding these mocks fairly dificult to reason about when trying to make this test pass, so decided to lift most of the mocks out of static methods and into a setup block (mostly because that is what I've seen Plumpy do when creating mocks, but also because it became a bit easier to see which mocks relied on which other mocks, and which were actually unnecessary). Let me know if you think this actually just made things more confusing...

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>