feat(google): add predictive autoscaling support (#8495)
* feat(google): add feature flag for predictive autoscaling

* refactor(google): move autoscaling policy interfaces and layout components to autoscalingPolicy directory

Currently, the only Typescript/React components of the autoscaling policy modal are specific to the scale-in control feature. In the next commit, we will add support for predictive autoscaling, so in order to share these interfaces and layout components, lift them to the top level.

* feat(google): add predictive autoscaling support (add/edit policy modal)

* feat(google): add predictive autoscaling support (SG details view)

* fix(google): sp/scaleDownInConfigured/scaleInControlsConfigured

I think we missed one spot when we renamed scale-in to scale-down.

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>