perf(kubernetes): Precompile find path for artifact replacers (#4801)
* test(kubernetes): Add some tests on ArtifactProvider and Replacers

Let's add a few tests on the ArtifactProvider. As it depends on an
implicit contract with the caching agent, let's use the integration
test rather than mock a lot of CacheData items that look like how
the caching agent currently stores data.

In a recent PR, we converted the ArtifactReplacer tests to Java and
also improved the test coverage. While these covered the general
interface of ArtifactReplacer by using a few specific replacers,
we still don't have good test coverage of all of the replacers
defined in Replacer.java.

This is important as the functionality of these replacers is almost
entirely encoded by complex JsonPath expressions, and I'd like to
refactor these to be cleaner.

This commit adds a ReplacerTest class that goes through each defined
replacer and tries to use it to find and replace artifacts. This should
provide sufficient coverage that we can refactor these replacers with
confidence.

* refactor(kubernetes): Push account filtering to artifact provider

All consumers of the artifact provider immediately filter the
results to a particular account; let's push this logic to the
artifact provider so they don't need to.

This removes the need for the checkArtifactAcrossAccount test,
which ensures that the versioned artifact converter, as it is
now replaced by a test on the artifact provider itself to ensure
that it only returns artifacts for a single account.

* refactor(kubernetes): Use filters in JsonPath expressions

The Replacer class contains two compilcated strings representing a
JsonPath; in all cases these are almost identical. Let's replace
this logic with something a bit more clear.  Ideally we'd not need
to use JsonPath because we would have a strongly-typed object and
would be able to use that, but for now let's at least improve this
code (and open up some performance improvements to come).

First, findPath and replacePath are always of the exact same form;
we build a query to find certain nodes. The findPath grabs all of
them while the replacePath filters them to ones matching a supplied
artifact. In addition to duplicating logic, the syntax of filter
conditions is very hard to read.

The JsonPath library allows you to put a placeholder [?] to represent
a filter condition, then bind it to a supplied object representing
the condition. This is exactly what we want; we want the same query
in all cases but to bind to "match everything" in the find case and
to bind to "match this artifact" in the replace case.

Let's require the replacers to supply a path with a [?] placeholder,
and then to supply a filter to determine whether to match a particular
artifact. There are two replacers that also require filtering for the
find case, so we'll allow this as an optional field (but will default
to finding everything).

Also add a NonnullByDefault to the class which replaces the need for
some individual Nonnull annotations.

* perf(kubernetes): Precompile find path for artifact replacers

The find path for artifact replacers is completely known a the time
the object is constructed. Let's pre-compile it so that we don't need
to do this compilation every time we look up artifats (which is every
time we return a manfiest from the /manifests endpoint).

We can't do the same with the replace query as it depends on the
input artifact. This isn't as much of an issue as we only replace
during a deploy rather than any time you request an existing manifest.
But let's slightly simplify the API by having the constructor handle
composing the find/replace queries.

* refactor(kubernetes): Clean up filterArtifacts function

Rename the verbose filterKubernetesArtifactsByNamespaceAndAccount
to simply filterArtifacts.

To make the logic more clear split it out into a few predicates
and use those predicates in the filter logic.

Finally, require namespace to be non-null and push that requirement
outward. In one case there's something we don't know for sure is
non-null so convert it with nullToEmpty.

* refactor(kubernetes): Simplify findGreatestUnusedVersion

There are a few simplifications to make here:
(1) We don't need to collect all the prior version to a sorted
array then grab the last element; we can just use max() to find the
maximum used version. This is slightly more efficient, O(N) vs
O(N log N) but really just makes the code simpler.
(2) We don't need to special-case the case of a version >= 1000;
the format string v%03d means pad to a *minimum* width of 3, and
handles numbers greater than 999 just fine.

Let's also add a couple of tests to the already good test coverage
to be sure.

* refactor(kubernetes): Clean up exception handling

We don't need to log right before throwing an exception; a log
message is only useful when suppressing the exception, as it's going
to show up in the logs anyway.

When catching an IOException wrap it in an UncheckedIOException rather
than a RuntimeException.

Finally, don't catch a generic Exception, instead catch a RuntimeException.

* refactor(kubernetes): Inline some variables

In both implementations of toArtifact we store a bunch of fields
from a KubernetesManifest to local variables in order to use them
a few lines later. This just adds more indirection and is confusing;
just read the fields we need from the KubernetesManifest.

Also getVersion should just take the KubernetesManifest itself;
it's just confusing to extract a bunch of fields into a long list
of arguments instead of just passing the manifest itself.

* refactor(kubernetes): Minor cleanup of find/replace logic

The find function can return a Stream as it's only used in one
place and the collection it returns is immediately converted to
a Stream.

Minorly refactor the replace function as well.  The goal here is
just to make it more clear which functions have side effects by
not having them deeply nested in other functions.

* refactor(kubernetes): Rename getType to getArtifactType

Also it should just accept a kind rather than a full manifest, as
it only needs the kind.

* refactor(kubernetes): Clean up nameFromReference

This is really minor, but the function is hard to read because it
switches from checking index >=0 to index < 0 partway through; let's
make it symmetric.

Also make the class final; it has a private constructor in any case
so it already is effectively final.

* Update clouddriver-kubernetes/src/test/java/com/netflix/spinnaker/clouddriver/kubernetes/artifact/ReplacerTest.java

Co-authored-by: Maggie Neterval <mneterval@google.com>

Co-authored-by: Maggie Neterval <mneterval@google.com>