fix(cats/sql): Fixes CloudDriver SQL Cache from failing due to id or agent being too long to be stored.  (#4755)
* fix(cats/sql): Reverting temporary fix when agent is too long for SQL table.

This temporary fix can still have issues when the type in the agent is longer than 127 characters, so I'm reverting it in favor of a proper long-term fix. This reverts commit fae18a1ad60582210a8158717aef248698ec817e.

* fix(cats/sql): Adding cats v2 template tables.

* fix(cats/sql): Fixes CloudDriver SQL Cache from failing due to id or agent being too long to be stored.

Fixes github.com/spinnaker/spinnaker/issues/5600. The theoretical max length of an id or agent is much larger than we can store in columns that are primary or index keys. Therefore, in order to solve the issue I moved away from storing the id and agent in key columns. I've introduced new columns that will be used as keys that store hashes of the ids and agents instead. I made no changes to the contracts between the SQL Cache and those who depend on it, all the changes are internal to the SQL Cache.

I also took the liberty of adding more javadoc style comments to some of the functions in the SQL Cache. Even though some of them are simple, I still think it's beneficial to introduce it since if the functions do become more complex then whoever updates them will need to update the comment.

* fix(cats/sql): Add drop tables by version to CATS SQL Admin Controller.

* Address first set of PR comments.

* Fix Kotlin spotless check issues

* Fix NPE due to null ids, move persistence logic out of admin controller, and prevent harmful SQL injection in sql admin controller.

* Add id index on cats v2 resource template. Will unfortunately limit id to 768 characters, but is necessary to avoid performance regression on the SqlCache filterIdentifiers function.

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>