perf(kubernetes): Reduce latency of streaming job executor (#4804)
* perf(kubernetes): Reduce latency of streaming job executor

Now that we always live lookup events when loading a manifest, the
latency of that call becomes more important. (It was always important
for users with liveManifestCalls, but now becomes important for
everyone.)

In general, for kubectl calls that might return multiple items, we
parse the output as it is produced to avoid needing to buffer a lot
of intermediate data. We do this by requesting asynchronous execution
of the job, then using the main thread to parse any output from the
job as it is produced.

It turns out that the apache-commons library we use adds a fair amount
of latency for quick asynchronous jobs. This is because the .waitFor()
function on the resultHandler just does a poll/sleep loop to wait for
the forked process to finish, which causes us to often wait quite a
bit longer than necessary. This matters much less for long jobs (or
for background processes) but for responding to user requests and in
cases where the call is fast (which is usually the case for getting
events where there are often none) this is significant.

Instead of requesting the job to run asynchronously and using the
main thread to parse the output, let's do the opposite. We'll send
an asynchronous request to parse the output, then run the job
synchronously on the main thread. This avoids the issues with latency
because none of these threads poll.

Ideally we'd move away from this commons-exec library and just use
ProcessBuilder, but given that we'll soon move clouddriver-kubernetes
to use the client library, it's not worth the effort at this point.

In my testing, this reduced the average time for a call to get
manifests from 268ms to 210ms (a 22% reduction); the reduction of
~50ms roughly matches the polling freqency from before, which makes
sense as now we don't need to wait a polling cycle to return.

* style(core): Rename executor to executorService

This avoids shadowing it in the function where it is used.