perf(kubernetes): Load manifest and instance details in parallel (#8499)
* refactor(kubernetes): Flip an if condition

retrieveInstance has a really long if block that would be clearer
if we just returned null immediately if the data is missing (rather
than have the usual path in a long block). Make that change separately
to make the next diff smaller.

* perf(kubernetes): Load manifest and instance details in parallel

When opening an instance tab, we currently wait for the reqeust to
/instances to finish before starting the request to /manifests even
though we have the required information for both.

This PR updates the Kubernetes instance provider to send both requests
immediately, reducing the amount of time the user needs to wait for
the instance tab to load.