perf(kubernetes): Request metrics and events in parallel (#4807)
* perf(kubernetes): Only deserialize pod status

Rather than convert the entire manifest to a V1Pod, only convert the
status as that's all this function is using.

Ideally we'd only do a single conversion to the full appropriate type
when we get a value back from the cluster, and pass that around, but
as long as we're still doing these ad-hoc intercoversions, it's worth
avoiding doing more than we need.

* perf(kubernetes): Request metrics and events in parallel

When deck requests a manifest, it wants the events and metrics as
well. This means that we make 3 sequential calls to the cluster,
each with a bit of latency, that ends up making the overall call
longer than needed.

Let's speed this up by sending the event and metrics requests
to an executor pool that runs them on a separate thread. This
allows us to only wait as long as the longest request instead
of the sum of all of the request times.

* fix(kubernetes): Add name to threads in executor pools

Add names to the the threads added in this PR and my last one.

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>