refactor(kubernetes): Remove some un-needed code (#4813)
* refactor(core): Add default implementations of deprecated methods

There are a few methods in the models that are deprecated in favor
of calling a different method. But because these don't have default
implementations, every implementation needs to override it which
adds more and more calls to these methods.

This commit adds default implementations that point to the recommended
method. I've removed the now-redundant override from the Kubernetes
implementations, but will let the maintainers of other providers remove
their overrides if they would like.

* refactor(kubernetes): Remove needless null checks on handlers

A while back I made getHandler always return non-null, but didn't
update all the place where we unnecessarily do this null check.

This commit updates the remaining places so I don't clutter up all
my other PRs with doing these one-by-one.

Also do two things where it makes things clearer:
* Remove the intermediate KubernetesResourceProperties variable
if we're only using it to look up the Handler.
* Move the assignment of the Handler to right before it is used
(and even inline it if it's only used once)

* refactor(kubernetes): Push multi-manifest description down

The KubernetesMultiManifestOperationDescription class is only
used by the delete operation; let's push it down and figure out
what to extract to a base class if/when we need another
implementation.

* refactor(kubernetes): Delete some deprecated functions

These functions all have unsafe unchecked casts; 6 months ago I
moved the only callers to convert to a strongly-typed object and
use the accessors on that object and deprecated these functions.
Let's remove them now.

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>