refactor(kubernetes): Enable error-prone and compiler warnings (#4819)
* fix(kubernetes): Fix errorprone TypeParameterUnusedInFormals

[TypeParameterUnusedInFormals] Declaring a type parameter that is
only used in the return type is a misuse of generics: operations on
the type parameter are unchecked, it hides unsafe casts at invocations
of the method, and it interacts badly with method overload resolution.

This check is 100% correct here; the only advantage of using the generic
here is that we can hide the casting in getRequiredField. In fact, let's
just get rid of this method and force callers to do the casting; keeping
the method just to throw the exception seems more confusing than inlining.

* fix(kubernetes): Fix error-prone issues

This fixes a few error-prone issues that were trivial to fix, and
were easier than suppressing the whole warning class.

[EmptyBlockTag] A block tag (@param, @return, @throws, @deprecated)
has an empty description. Block tags without descriptions don't add
much value for future readers of the code; consider removing the tag
entirely or adding a description.

In this case just delete the tag as the name is enough.

[UnusedNestedClass] This nested class is unused, and can be removed.

We were indeed not using this nested class.

[PublicConstructorForAbstractClass] Constructors of on an abstract class
can be declared protected as there is never a need for them to be public.

[UnnecessaryLambda] Returning a lambda from a helper method or saving
it in a constant is unnecessary; prefer to implement the functional
interface method directly and use a method reference instead.

Turns out we weren't using this lambda saved to a variable at all.

[UnusedMethod] Private method is never used.

error-prone doesn't seem to be able to detect that JUnit is using these
as a MethodSource for parametrized tests; just suppress the warnings here.

Finally, this isn't from error-prone, but there's a compiler warning that
KubernetesNamedAccountCredentials overrides equals but doesn't call the
super class. We don't want to do that here (because there's nothing in
the super class and because we only want some explicitly-included fields)
so add callSuper = false to stop the warning.

* refactor(kubernetes): Enable compilation warnings

This commit enables compilation warnings for clouddriver-kubernetes.
It turns off a few categories that either generate too many false
positives for our code (with a detailed comment) or that have too
many actual issues we need to fix first.

The plan is to fix the categories of issues one-by-one and
re-enable them.

* feat(kubernetes): Add error-prone to clouddriver-kubernetes

This commit adds the error-prone static analysis checker to
clouddriver-kubernetes. It disables any of the checks that have
actual issues.

The plan is that we can fix the issues in a given category then
enable that category, until all of the issues are fixed.

* fix(kubernetes): Add comment to disabled checks

Let's add a comment about why the checks are disabled.