refactor(kubernetes): Fix a few warning types (#4823)
* refactor(kubernetes): Apply type bounds to implementations of AtomicOperation

Now that the interfaces have type bounds, we can apply these to the implementations
as well.

* refactor(kubernetes): Suppress some unchecked warnings

KubernetesManifest is full of unchecked casts; let's just add
a comment to the class and suppress the warnings for any methods
that currently have unchecked casts.

As noted in the comment, we should really try to stop relying on
this class that extends HashMap and is full of these unsafe casts,
but for now we'll just suppress the warnings.

I have not enabled the Xlint:unchecked flag because there are still
a few other unchecked warnings that I'll leave for later.

Also let's make computedKind transient because the class implements
Serializable. It doesn't really matter because we don't use java
serialization, but is more correct because this is just a cached
value that is stored to avoid re-computing.

* fix(kubernetes): Fix deprecation warnings

This commit fixes the small number of deprecation warnings in
clouddriver-kubernetes and enables deprecation compiler warnings.

Notes:
* getPointCoordinates was marked deprecated, but is still actively
used in the Delete Manifest stage; there appers to have been no
effort to actually deprecate it, so let's just remove the annotation.
* Stop logging the full manifest when we hit the warning on
that deprecated call; just the identifying info should be enough.
* getRequiredGroupMembership is a field; to add the suppress
annotation we need to write the lombok-generated accessor
explicitly

Also two small other fixes I noticed:
* Make a field private
* Remove Serializable from KubernetesV2Job status as we never use
Java serialization.

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>