fix(*): Fix gradle deprecation warnings (#4826)
* fix(*): Fix gradle deprecation warnings

Because I'm in a warning-fixing mood this week...this fixes the
warning when running builds that deprecated features are used.

These were all pretty easy to fix:
* Instead of force, need to use version constraints
* Instead of testCompile/testRuntime, need to use testCompileOnly/
testRuntimeOnly.

* fix(elasticsearch): We need containers at test runtime not just compile

* fix(*): Pin elasticsearch in clouddriver-web

The new gradle rich versions feature is requiring us to be explicit
about dependency conflicts (whereas before it would resolve things).

Given that clouddriver-elasticsearch and clouddriver-elasticsearch-aws
pin different versions of elasticsearch, including them both in
clouddriver-web is causing a build failure.

For now I've resolved this by just picking the version that was
resolved before explicitly, but we probably should try to get these
on the same version.

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>