fix(kubernetes): Do a live lookup in the artifact provider (#4831)
* refactor(kubernetes): Update getArtifacts to accept kind

As we'll be directly querying the cluster, it makes more sense to
accept the actual kind rather than the artifact kind, which is
"kubernetes/kind". For now we'll just re-construct the kind in the
function, but that will shortly be unnecessary when we move to
directly queryin the cluster.

* fix(kubernetes): Do a live lookup in the artifact provider

This is only used during deploys, so latency is not a big issue
here. This will fix race conditions with live manifest calls
enabled where operations later in a pipeline don't account for
operations that happened earlier in the pipeline.

For now I am just replacing the exact same logic with a live call,
including:
* Building the whole artifact (even though we only need a few fields)
* Sorting by creation timestamp instead of by version

This unblocks deleting the code that caches artifacts, which then
unblocks some futher simplifications.

* test(kubernetes): Move a test to Java

As part of my goal to incrementally move tests to Java, if I break
a test and need to fix it, I'm converting it to Java. The next commit
breaks this test, so first convert it to Java.

* refactor(kubernetes): Stop caching artifacts

Now that we always do a live lookup of artifacts, we can stop
caching them, as no code reads artifacts from the cache.

I've left the enum constant for artifact cache keys, and left
the core caching agent authoritative, so that existing entries
get evicted. We can remove this code after we've done a release
with it present.