feat(*): Add optional code coverage reports (#4832)
This commit applies the jacoco pluigin to all subprojects,
giving them the task jacocoTestReport to get a test coverage
report.

When the plugin is present, tests are instrumented by default
(as the tests don't know if we'll later want to generate a
report), which slows them down. To prevent slowing down all
builds, I've put the test instrumentation behind a flag.

Now if you run
./gradlew -PtestCoverage jacocoTestReport
it will instrument the tests and generate a coverage report
for each submodule.

Also configure lombok to add a Generated annotation to code that
it generates, so that it gets excluded from code coverage reports.

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>