feat(kubernetes): DeployManifestStage makes live calls to determine cluster members for rollout strategies  (#3879)
* feat(kubernetes): DeployManifestStage makes live calls to determine cluster members for rollout strategies

In order to make the `liveManifestCalls` behavior the default for all users, we must change all pipeline operations that rely on cache consistency to instead call Clouddriver endpoints that make live queries against the Kubernetes cluster.

This commit replaces the existing `getClusters` call in the DeployManifestStage, used to hydrate the context of any Disable and Delete stages necessary in order to enact a Red/Black or Highlander rollout strategy, with a new `getClusterManifests` call, which returns a list of ManifestCoordinates for members of a given cluster. Also adds tests asserting on the basic correctness of the stage graph when using rollout strategies.

* refactor(kubernetes): remove raw context Maps in DeployManifestStage

Leaving the other non-context raw Maps as-is because those are all a result of deeply inspecting the list of full manifests currently present in the stage context as raw Maps, which we'll replace in an upcoming PR with a list of ManifestCoordinates.

* test(kubernetes): make DeployManifestStageTest more realistic

The just-deployed resource should be returned among the cluster manifests fetched.

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>