feat(kubernetes): add endpoint to list cluster manifest coordinates  (#4841)
* feat(kubernetes): add endopint to list cluster manifest coordinates

This will be consumed by Orca's Deploy Manifest Stage when fetching SGs to disable/delete as part of rollout strategy workflows.

* refactor(kubernetes): minimize accessibility of ManifestController and KubernetesManifestProvider

* refactor(kubernetes): throw error if account unresolvable

The only situation getClusterManifestCoordinates returns null in is when we can't resolve an account name. Let's just throw an error as soon as possible instead of throwing a less meaningful error downstream.

* refactor(kubernetes): make symmetrical changes to getClusterAndSortAscending

* fix(kubernetes): only consider same-app clusters for dynamic target selection

This preserves symmetry with the rollout strategies logic and consistency with the behavior when this endpoint used the cache instead of making live calls.

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>