refactor(kubernetes): Fix JdkObsolete warnings (#4850)
* refactor(kubernetes): Minor refactor of timestamp function

Instead of looking twice in the map, just check if the result is
null in getCreationTimestamp.  (Since this is a mutable map, the
presence of the key doesn't guarantee the value is null anyway, so
let's jut get the value.) I didn't use Optional here because this
is used in a comparator that might be called a lot for large sorts,
but maybe that's a micro-optimization.

The result is non-null so we can use .isEmpty() in getFormattedTimestamp.

Finally let's rename getFormattedTimestamp to getCreationTimestampEpochMillis,
as the result is not formatted.

* refactor(kubernetes): Fix JdkObsolete warnings

[JdkObsolete] Date has a bad API that leads to bugs; prefer
java.time.Instant or LocalDate.

In KubernetesManifest, we can parse the String into an Instant
instead of a Date. (There's also no point separately handling
an empty string; anything invalid will cause a ParseException
so just let that happen.)

In KubernetesModelUtil, the only place that was calling this
function already had a org.joda.time.DateTime object and was
serializing it before calling this function; we can just directly
get the milliseconds from the object and delete translateTime.