fix(kubernetes): only workload types have cluster relationships (#4861)
* refactor(kubernetes): move hasClusterRelationship from KubernetesKindProperties to KubernetesCacheDataConverter

Moves logic determining whether to cache cluster relationships for a given kind to the only place it is needed. (Will also make next commits more readable.)

* fix(kubernetes): differentiate between logical and cluster relationships

Only kinds that map to Spinnaker server groups and server group managers are considered Spinnaker clusters, and should be differentiated between the other "logical" kinds surfaced in the Application UI as Spinnaker Load Balancers and Firewalls. Stop caching cluster relationships for non-cluster kinds, and update tests accordingly.

* refactor(kubernetes): make logical and cluster relationships agnostic to Kubernetes kind

We map Kubernetes kinds to Spinnaker kinds in order to surface them in the Application UI as Clusters, Load Balancers, and Firewalls. Looking ahead to when CRDs may be mappable to Spinnaker kinds as well, let's not keep a fragile list of Kubernetes kinds that have these Spinnaker-kind-specific relationships, and instead just check by Spinnaker kind.

(This commit is not necessary to solve the problem this PR set out to solve, so if y'all think this actually makes it less clear or it's too cumbersome to pass around the SpinnakerKubernetesKindMap, happy to remove it from this PR!)

* refactor(kubernetes): remove unnecessary getters; use ImmutableSets instead of ImmutableLists

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>