feat(logging): Ensure that request duration reflects the entire request (#1326)
The `RequestLoggingFilter` has always run after the security filter chain
_and_ the `AuthenticatedRequestFilter`.

This unfortunately had a side-effect of the tracked duration not reflecting
time spent during the authentication flow.

By introducing a new higher priority filter, we can capture an accurate
start time for a request as well as have a request id that can be
propagated throughout the entire lifetime of a request. The latter is
a helpful debugging technique.