feat(api): allow fiat clients to check if there is a user permission cached (#768)
For sessionless auth such as X509 in gate, there is a debounce option to prevent having to post a login to fiat for each request,
however part of that is determining whether the user is already populated in the hosts permission cache.

This check 404s if the user has not logged in, but in doing so triggers retries with backoffs before a login is posted.

This introduces a new method just to examine the cache without calling through to the loading behaviour on a miss