fix(kubernetes/artifacts): Update Clouddriver's artifact replacement logic to replace docker images even if the image specified in the manifest includes a tag. (#4874)
* fix(kubernetes/artifacts): Update Clouddriver's artifact replacement logic to replace docker images even if the image specified in the manifest includes a tag.

Previous implementation didn't allow overwriting a image in a manifest with a bound artifact if the input manifest had a tag on it. This was problematic when using helm since most helm charts have a tag.

BREAKING CHANGE: Can no longer rely on Spinnaker ignoring any matching image artifacts available in the pipeline and using the tag specified in the manifest image.

* Use nameFromReference to extract the artifact name in the input manifest rather than using regex.

* Use JsonPath in replace filter predicate.

* Change artifact replacement in KubernetesDeployManifestOperation from replacing required then optional artifacts to replacing optional then required artifacts. This change is being done to prevent optional artifacts from replacing required artifacts.

* String check to avoid potential exception.

* Move String check in replace filter predicate to avoid potential exception.