fix(actuator): broaden scope of try block (#1576)
* fix(actuator): broaden scope of try block

We have some resource checks that are failing because of a failed
`checkNotNull` check. However, these failures weren't incrementing
our metric that tracks error checks.

The check happened outside of the try block, so the relevant
catch block never happened.

This change widens the scope of the try block to ensure that
a failed check gets properly captured as an error.

* fix(pr): add back zapped carriage return