perf(cats/sql): Introduce V2 SQL cache schema (#4998)
* fix(cats/sql): Reintroduce fix to Clouddriver SQL Cache failing due to id or agent being too long to be stored.

Revert "fix(cats/sql): Reverting changes to SQL Cache due to seeing a major increase in CPU usage and deadlocking. (#4811)"

This reverts commit 618921b0d0f8f9f1aed30e16a8f44abff6348acd.

* perf(cats/sql): Change relationship table index with rel_agent_hash to include rel_type.

Including the rel_type in the the index will improve the query where rel_agent_hash and rel_type is given. There's also a query to get all entries with a given rel_agent_hash, this index still works for that query since the rel_agent_hash is the first column.

* perf(cats/sql): Switch SqlCache hashing function from SHA256 to Murmur3.

Moving towards using a faster hashing function since we're spending a lot of CPU time hashing at the moment. Furthermore, the values we are hashing do not need to be hashed by a cryptographic hash function so SHA256 is overkill.

* fix(cats): Fix regressions in nullable identifiers

Co-authored-by: Matt Santa <mattsanta@google.com>
Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>