chore(build): change Dockerfile.compile base to openjdk:11-slim (#8652)
The move to `alpine` caused build failures when attempting to use the
`node` binary provided by the `nebula.node` Gradle plugin. The binary
was built assuming the presence of `glibc` while Alpine is based on
`musl`, so the build would always fail.

Moving to openjdk:11-slim appears to get us around these build issues,
we'll see in the final build in gcloud if that's correct!