fix(export): handle optional fields in context (#1663)
When attempting to export a cluster who was deployed by a pipeline
whose deploy stage context is missing `delayBeforeDisableSec` or
`delayBeforeScaleDownSec`, keel would throw a 500 error with
the following stack trace:

```
java.lang.NumberFormatException: For input string: "null"
  at java.lang.NumberFormatException.forInputString(NumberFormatException.java:65)
  at java.lang.Integer.parseInt(Integer.java:652)
  at java.lang.Integer.parseInt(Integer.java:770)
  at com.netflix.spinnaker.keel.orca.ClusterExportHelper.contextToRedBlack(ClusterExportHelper.kt:134)
  at com.netflix.spinnaker.keel.orca.ClusterExportHelper.access$contextToRedBlack(ClusterExportHelper.kt:16)
```

Since the `RedBlack` class already supported these two fields as
nullable, return null if they are not present.

This is a follow-up to #1661, which handled one of the optional fields.
There were two other ones that were missed. Only this time, there are
tests!

You'll notice that the code does this:

```kotlin
delayBeforeDisable = this["delayBeforeDisableSec"]?.toString()?.toInt()?.toLong()?.let{ Duration.ofSeconds(it) },
```

instead of the simpler:

```kotlin
delayBeforeDisable = this["delayBeforeDisableSec"]?.toLong()?.let{ Duration.ofSeconds(it) },
```

That's how the code was initially, and so I left it.

I asked @robfletcher, who git blame suggested was the original author, and he
couldn't remember the specific reason, but had a vague sense that there
was one. So I decided to leave it.